# Add scene entity memoization to improve converter function efficiency