This draft PR serves as a first benchmark/demonstration on the impact of caching scene entities in a weak map instead of re-generating in every call of the converter function for repeated calls for the same OSI frame. The caching mechanism aims to minimize/compensate the redundant frame loading overhead when opening an OSI mcap file (see https://github.com/Lichtblick-Suite/lichtblick/issues/382).

There are two functions (`convertGroundTruthToSceneUpdateCache` and `convertGroundTruthToSceneUpdateNoCache` = current implementation of 0.0.4) in this PR. Each can be referenced in `registerMessageConverter` for comparison and performance evaluation. The performance of each option can be measured by aggregating the total time spent in the converter function until lichtblick stops calling it after just opening the file.

On my machine, these are the results for the `20250110T150000Z_sv_3.7.0_highway_drop-bike_5ms.mcap` (see [here](https://github.com/thomassedlmayer/lichtblick-extension-bug-demo/blob/master/20250110T150000Z_sv_3.7.0_highway_drop-bike_5ms.mcap)) example file until the function calls finally stop after about 41.000 calls:
- `convertGroundTruthToSceneUpdateCache`:
![image](https://github.com/user-attachments/assets/9f1337d1-c94e-4cca-8a6f-94e51ce924fe)
Running this experiment multiple times (~10) yielded values between 500 and 1000ms.

- `convertGroundTruthToSceneUpdateNoCache`:
![image](https://github.com/user-attachments/assets/4415f5d0-b9a3-4eb4-8b36-5e80511f2a9b)
Running this experiment multiple times (~10) yielded values between 10000 and 15000ms.

To reproduce the benchmark, the timeline of lichtblick should not be clicked until the calls/logs stop. The results above show the final console info log after opening the file and not clicking anything.

Note that the measured times (~0.6s / ~15.6s) do not represent the total time until the "file loading bug" stops calling, it's just the time that is spent inside the converter function. So this does not reduce the total loading time down to 0.6s. It still takes some more seconds until the calls are "done". But that's about all we can do in terms of performance optimization in the extension.

The general advantage of this change is that the initial loading time of a file is compensated and that lichtblick stops lagging at least a bit earlier. Unfortunately, it does not show significant improvements when playing a file - at least on my machine. It also does not resolve the "full cache issue" (see https://github.com/Lichtblick-Suite/lichtblick/issues/383).