@jdsika 2025-02-21
Did you notice anything specific about the memory consumed comparing cached vs. non-cached? Otherwise I am having a hard time to imagine any bad effects of this approach?... should we maintain the time logging when merging and maybe disable it with a compiler flag in case we need it again?
(I will play around with it on Monday)

I am curious about your assessment regarding this comment:

> Unfortunately, it does not show significant improvements when playing a file

What do you suggest as next step? Check out the physical lane algorithm and test correctness of each element? OR would you go for the "live time" and think about #78 ? I am not sure if the converter is even allowed to skip the calculation of messages...