@jdsika 2025-02-21
Another idea could be to circumvent the cache "pre-loading" issue by using the websocket to stream the messages to the application? See if it changes behavior: https://github.com/Lichtblick-Suite/ws-protocol/tree/main