@thomassedlmayer 2025-02-25
> Did you notice anything specific about the memory consumed comparing cached vs. non-cached? Otherwise I am having a hard time to imagine any bad effects of this approach?...

With caching the memory utilization is slightly lower during initial file loading (right after opening the file). When playing the file, the memory utilization peaks are a bit higher (a few 100MB) as compared to non-caching but it also frees up again after some time when pausing. I'd say the impact is not really noticeable in both situations.

I also did some more testing: Without caching it takes about 55 to 60 seconds for the CPU load to go down to 0% after opening the example file whereas it takes about 25 to 30 seconds with caching.

I'm still having issues after reloading lichtblick (Ctrl+R) and reopening a file. The memory utilization stays quite high after the reload and then rises even more when opening the file again. Lichtblick usually crashes after some minutes of being completely unusable and laggy (with CPU utilization of >30%). It seems like it does not free up the memory when doing a reload? A "hard" restart of the app is necessary for me to have acceptable performance; but this applies for both caching/non-caching options. It does not happen with other really big mcap files (>18GB; with lidar data etc.) that don't require an extension to be visualized. I can repeatedly reload lichtblick and reopen different big non-OSI mcap files without issues. Might be an issue of the extension integration in lichtblick? Though, playback seems to be similarly "fast".

> I am curious about your assessment regarding this comment:
"Unfortunately, it does not show significant improvements when playing a file"

My guess would be that the rendering process (or whatever lichtblick does) takes so much time that the extension performance does not really matter when playing a file and viewing it in the 3d viewer. For initial file loading it probably only improves the performance because the scene updates are just ran through without being actually rendered in the viewer.

> should we maintain the time logging when merging and maybe disable it with a compiler flag in case we need it again?

I don't know how much the extension performance really matters as it most probably isn't the bottleneck (0.5s total spent time in the extension vs. 25s total initial loading time). But of course, for testing it sometimes might be useful to check if you have messed up something. I'd do that in a separate PR, but give it less priority for now.

> Another idea could be to circumvent the cache "pre-loading" issue by using the websocket to stream the messages to the application? See if it changes behavior: https://github.com/Lichtblick-Suite/ws-protocol/tree/main

I will try to have a look at this as soon as the lanes are implemented.

> What do you suggest as next step? Check out the physical lane algorithm and test correctness of each element? OR would you go for the "live time" and think about https://github.com/Lichtblick-Suite/asam-osi-converter/issues/78 ? I am not sure if the converter is even allowed to skip the calculation of messages...

I'm currently working on the lane boundary implementation. We can then log the lichtblick scene entity objects for every step (containing all vertices for the lane boundaries; not boundary points but actual vertices of the triangles that are being rendered) and use that data in our validation framework for validation against a reference implementation.

For checking the "determineTheNeedtoRerender" function and static objects cache we need some files with higher frame frequency (<0.1s step size) and traffic signs (since both mechanisms are applied only to traffic signs; I don't really know why tbh). We will have to create some example files that comply with those requirements first. Or do we already have some?