@jdsika 2025-02-25
![disable_pre-load](https://github.com/user-attachments/assets/83554e0b-d3d1-468b-9143-e6215a5c0ab3)

Apparently there is an option to disable pre-load