@thomassedlmayer 2025-02-25
Well, this solves a lot of performance issues... Didn't see that yet.

As far as I understood the pre-loading feature is supposed to ensure correctly rendered frame transforms if you use accumulating frame transforms (see https://foxglove.dev/blog/improving-how-foxglove-loads-data). But we don't have that in OSI because every frame must be unambiguously interpretable on its own anyways.